+++ src/Git_Updater/Traits/API_Common.php
@@ -72,8 +72,12 @@ trait API_Common {
 			} else {
 				$release_assets = [];
 				foreach ( $response as $release ) {
-					$release_assets[ $release->tag_name ] = $release->assets[0]->url ?? '';
+					// Ignore leading 'v' and skip anything with dash or words.
+					if ( ! preg_match( '/[^v]+[-a-z]+/', $release->tag_name ) ) {
+						$release_assets[ $release->tag_name ] = $release->assets[0]->url ?? '';
+					}
 				}
+				uksort( $release_assets, fn ( $a, $b ) => version_compare( ltrim( $b, 'v' ), ltrim( $a, 'v' ) ) );
 
 				return $release_assets;
 			}